move cam help dialog to intersections
