+++ web-app/src/app/client/modules/intersection/cam-help-dialog/cam-help-dialog.component.ts
